+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 ## Unreleased 0.59.0
 
 ### 🚨 Breaking changes
+- [6505](https://github.com/vegaprotocol/vega/issues/6505) - Allow negative position decimal places for market
 - [6477](https://github.com/vegaprotocol/vega/issues/6477) - Allow the user to specify a different passphrase when isolating a key
-
 
 ### 🗑️ Deprecation
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -